add role-based authorization; introduce ADMIN role and set default user role to RENTER; refactor property and user routes to enforce role checks
