+++ prisma/schema.prisma
@@ -30,7 +30,7 @@ model User {
   name       String
   avatar     String
   email      String     @unique
-  role       Role
+  role      Role     @default(RENTER)
   verified   Boolean    @default(false)
   createdAt  DateTime   @default(now())
   updatedAt  DateTime   @updatedAt
@@ -93,6 +93,7 @@ model Booking {
  * User role enum type
  */
 enum Role {
+  ADMIN
   RENTER
   HOST
 }
