+++ src/types/request.types.ts
@@ -1,0 +1,6 @@
+import { IUser } from "./user.types";
+import { Request } from 'express';
+
+export interface CustomRequest extends Request {
+    user: IUser;
+}
